requirements.txt
supabase
fastapi[all]
python-dotenv
python-multipart
httpx
uvicorn
mysql-connector-python
pydantic